Rework fnvpair -- restore functionality and reduce differences
Upon reflection, I realized that replacing VERIFY0 with assert was a
mistake.  If assertions are disabled, the function will never be executed.
What we want is to execute the function in all cases, then verify the
result.  If the verification is disabled, we want functionality to remain.

I created a macro called VERIFY0 that does this.  Using it removes most
of the differences from Illumos that the previous commit introduced.  One
exception  is the "fnvlist_pack" function which used VERIFY3U instead of
VERIFY0.  I can't understand why the 3U version was being used (the result
is an integer like all the others) so I just used VERIFY0 here as well.